+++ node/src/RouterTypes.ts
@@ -30,7 +30,10 @@ import type {
 	AudioLevelObserver,
 	AudioLevelObserverOptions,
 } from './AudioLevelObserverTypes';
-import type { RtpCapabilities, RtpCodecCapability } from './rtpParametersTypes';
+import type {
+	RtpCapabilities,
+	RouterRtpCodecCapability,
+} from './rtpParametersTypes';
 import type { NumSctpStreams } from './sctpParametersTypes';
 import type { Either, AppData } from './types';
 
@@ -38,7 +41,7 @@ export type RouterOptions<RouterAppData extends AppData = AppData> = {
 	/**
 	 * Router media codecs.
 	 */
-	mediaCodecs?: RtpCodecCapability[];
+	mediaCodecs?: RouterRtpCodecCapability[];
 
 	/**
 	 * Custom application data.
@@ -61,6 +64,16 @@ export type PipeToRouterOptions = {
 	 * Target Router instance.
 	 */
 	router: Router;
+
+	/**
+	 * Whether the `id` of the returned Producer or DataProducer should be the
+	 * same than the `id` of the original Producer or DataProducer. Default true.
+	 *
+	 * @remarks
+	 * - If set to true, then the origin router and target router cannot be in the
+	 *   same worker (if so, `pipeToRouter()` with throw).
+	 */
+	keepId?: boolean;
 
 	/**
 	 * Create a SCTP association. Default true.
@@ -292,5 +305,5 @@ export interface Router<RouterAppData extends AppData = AppData>
 	 * Update the Router media codecs. Once called, the return value of the
 	 * router.rtpCapabilities getter changes.
 	 */
-	updateMediaCodecs(mediaCodecs: RtpCodecCapability[]): void;
+	updateMediaCodecs(mediaCodecs: RouterRtpCodecCapability[]): void;
 }