+++ node/src/rtpParametersTypes.ts
@@ -1,4 +1,9 @@
 /**
+ * Media kind ('audio' or 'video').
+ */
+export type MediaKind = 'audio' | 'video';
+
+/**
  * The RTP capabilities define what mediasoup or an endpoint can receive at
  * media level.
  */
@@ -15,9 +20,12 @@ export type RtpCapabilities = {
 };
 
 /**
- * Media kind ('audio' or 'video').
+ * Special RtpCapabilities for `supportedRtpCapabilities` in which `codecs`
+ * is an array of RouterRtpCodecCapability.
  */
-export type MediaKind = 'audio' | 'video';
+export type RouterRtpCapabilities = Omit<RtpCapabilities, 'codecs'> & {
+	codecs?: RouterRtpCodecCapability[];
+};
 
 /**
  * Provides information on the capabilities of a codec within the RTP
@@ -49,8 +57,11 @@ export type RtpCodecCapability = {
 
 	/**
 	 * The preferred RTP payload type.
+	 *
+	 * NOTE: Despite it's a mandatory field, it's optional in `mediaCodecs` of
+	 * RouterOptions.
 	 */
-	preferredPayloadType?: number;
+	preferredPayloadType: number;
 
 	/**
 	 * Codec clock rate expressed in Hertz.
@@ -74,6 +85,17 @@ export type RtpCodecCapability = {
 	 * Transport layer and codec-specific feedback messages for this codec.
 	 */
 	rtcpFeedback?: RtcpFeedback[];
+};
+
+/**
+ * Special RtpCodecCapability for RouterOptions in which `preferredPayloadType`
+ * is optional.
+ */
+export type RouterRtpCodecCapability = Omit<
+	RtpCodecCapability,
+	'preferredPayloadType'
+> & {
+	preferredPayloadType?: number;
 };
 
 /**