+++ rust/src/router/pipe_transport/tests.rs
@@ -2,7 +2,10 @@ use crate::consumer::ConsumerOptions;
 use crate::data_consumer::DataConsumerOptions;
 use crate::data_producer::DataProducerOptions;
 use crate::producer::ProducerOptions;
-use crate::router::{PipeToRouterOptions, Router, RouterOptions};
+use crate::router::{
+    PipeDataProducerToRouterPair, PipeProducerToRouterPair, PipeToRouterOptions, Router,
+    RouterOptions,
+};
 use crate::transport::Transport;
 use crate::webrtc_transport::{
     WebRtcTransport, WebRtcTransportListenInfos, WebRtcTransportOptions,
@@ -134,7 +137,10 @@ fn producer_close_is_transmitted_to_pipe_consumer() {
             .await
             .expect("Failed to produce video");
 
-        router1
+        let PipeProducerToRouterPair {
+            pipe_producer,
+            pipe_consumer: _,
+        } = router1
             .pipe_producer_to_router(
                 video_producer.id(),
                 PipeToRouterOptions::new(router2.clone()),
@@ -142,9 +148,11 @@ fn producer_close_is_transmitted_to_pipe_consumer() {
             .await
             .expect("Failed to pipe video producer to router");
 
+        let pipe_producer = pipe_producer.into_inner();
+
         let video_consumer = transport2
             .consume(ConsumerOptions::new(
-                video_producer.id(),
+                pipe_producer.id(),
                 consumer_device_capabilities(),
             ))
             .await
@@ -173,7 +181,10 @@ fn data_producer_close_is_transmitted_to_pipe_data_consumer() {
             .await
             .expect("Failed to produce data");
 
-        router1
+        let PipeDataProducerToRouterPair {
+            pipe_data_producer,
+            pipe_data_consumer: _,
+        } = router1
             .pipe_data_producer_to_router(
                 data_producer.id(),
                 PipeToRouterOptions::new(router2.clone()),
@@ -181,8 +192,10 @@ fn data_producer_close_is_transmitted_to_pipe_data_consumer() {
             .await
             .expect("Failed to pipe data producer to router");
 
+        let pipe_data_producer = pipe_data_producer.into_inner();
+
         let data_consumer = transport2
-            .consume_data(DataConsumerOptions::new_sctp(data_producer.id()))
+            .consume_data(DataConsumerOptions::new_sctp(pipe_data_producer.id()))
             .await
             .expect("Failed to create data consumer");