+++ rust/tests/integration/pipe_transport.rs
@@ -293,8 +293,8 @@ fn pipe_to_router_succeeds_with_audio() {
             .expect("Failed to produce audio");
 
         let PipeProducerToRouterPair {
-            pipe_consumer,
             pipe_producer,
+            pipe_consumer,
         } = router1
             .pipe_producer_to_router(
                 audio_producer.id(),
@@ -443,8 +443,8 @@ fn pipe_to_router_succeeds_with_video() {
             .expect("Failed to pause video producer");
 
         let PipeProducerToRouterPair {
-            pipe_consumer,
             pipe_producer,
+            pipe_consumer,
         } = router1
             .pipe_producer_to_router(
                 video_producer.id(),
@@ -562,7 +562,7 @@ fn pipe_to_router_succeeds_with_video() {
 }
 
 #[test]
-fn pipe_to_router_fails_if_both_routers_belong_to_the_same_worker() {
+fn pipe_to_router_with_keep_id_true_fails_if_both_routers_belong_to_the_same_worker() {
     future::block_on(async move {
         let (worker1, _worker2, router1, _router2, transport1, _transport2) = init().await;
 
@@ -591,6 +591,39 @@ fn pipe_to_router_fails_if_both_routers_belong_to_the_same_worker() {
         } else {
             panic!("Unexpected result: {result:?}");
         }
+    });
+}
+
+#[test]
+fn pipe_to_router_with_keep_id_false_does_not_fail_if_both_routers_belong_to_the_same_worker() {
+    future::block_on(async move {
+        let (worker1, _worker2, router1, _router2, transport1, _transport2) = init().await;
+
+        let router1bis = worker1
+            .create_router(RouterOptions::new(media_codecs()))
+            .await
+            .expect("Failed to create router");
+
+        let video_producer = transport1
+            .produce(video_producer_options())
+            .await
+            .expect("Failed to produce video");
+
+        let PipeProducerToRouterPair {
+            pipe_producer,
+            pipe_consumer: _,
+        } = router1
+            .pipe_producer_to_router(video_producer.id(), {
+                let mut options = PipeToRouterOptions::new(router1bis.clone());
+                options.keep_id = false;
+                options
+            })
+            .await
+            .expect("Failed to pipe producer to router");
+
+        let pipe_producer = pipe_producer.into_inner();
+
+        assert_ne!(pipe_producer.id(), video_producer.id());
     });
 }
 
@@ -879,7 +912,7 @@ fn consume_for_pipe_producer_succeeds() {
             .await
             .expect("Failed to pause video producer");
 
-        router1
+        let PipeProducerToRouterPair { pipe_producer, .. } = router1
             .pipe_producer_to_router(
                 video_producer.id(),
                 PipeToRouterOptions::new(router2.clone()),
@@ -887,9 +920,11 @@ fn consume_for_pipe_producer_succeeds() {
             .await
             .expect("Failed to pipe video producer to router");
 
+        let pipe_video_producer = pipe_producer.into_inner();
+
         let video_consumer = transport2
             .consume(ConsumerOptions::new(
-                video_producer.id(),
+                pipe_video_producer.id(),
                 consumer_device_capabilities(),
             ))
             .await
@@ -969,7 +1004,10 @@ fn producer_pause_resume_are_transmitted_to_pipe_consumer() {
             .await
             .expect("Failed to pause video producer");
 
-        router1
+        let PipeProducerToRouterPair {
+            pipe_producer,
+            pipe_consumer: _,
+        } = router1
             .pipe_producer_to_router(
                 video_producer.id(),
                 PipeToRouterOptions::new(router2.clone()),
@@ -977,9 +1015,11 @@ fn producer_pause_resume_are_transmitted_to_pipe_consumer() {
             .await
             .expect("Failed to pipe video producer to router");
 
+        let pipe_video_producer = pipe_producer.into_inner();
+
         let video_consumer = transport2
             .consume(ConsumerOptions::new(
-                video_producer.id(),
+                pipe_video_producer.id(),
                 consumer_device_capabilities(),
             ))
             .await
@@ -1044,8 +1084,8 @@ fn pipe_to_router_succeeds_with_data() {
             .expect("Failed to produce data");
 
         let PipeDataProducerToRouterPair {
-            pipe_data_consumer,
             pipe_data_producer,
+            pipe_data_consumer,
         } = router1
             .pipe_data_producer_to_router(
                 data_producer.id(),
@@ -1088,7 +1128,6 @@ fn pipe_to_router_succeeds_with_data() {
         assert_eq!(pipe_data_consumer.label().as_str(), "foo");
         assert_eq!(pipe_data_consumer.protocol().as_str(), "bar");
 
-        assert_eq!(pipe_data_producer.id(), data_producer.id());
         assert_eq!(pipe_data_producer.r#type(), DataProducerType::Sctp);
         {
             let sctp_stream_parameters = pipe_data_producer.sctp_stream_parameters();
@@ -1115,7 +1154,10 @@ fn data_consume_for_pipe_data_producer_succeeds() {
             .await
             .expect("Failed to produce data");
 
-        router1
+        let PipeDataProducerToRouterPair {
+            pipe_data_producer,
+            pipe_data_consumer: _,
+        } = router1
             .pipe_data_producer_to_router(
                 data_producer.id(),
                 PipeToRouterOptions::new(router2.clone()),
@@ -1123,8 +1165,10 @@ fn data_consume_for_pipe_data_producer_succeeds() {
             .await
             .expect("Failed to pipe data producer to router");
 
+        let pipe_data_producer = pipe_data_producer.into_inner();
+
         let data_consumer = transport2
-            .consume_data(DataConsumerOptions::new_sctp(data_producer.id()))
+            .consume_data(DataConsumerOptions::new_sctp(pipe_data_producer.id()))
             .await
             .expect("Failed to create data consumer");