+++ worker/src/RTC/SimpleConsumer.cpp
@@ -11,7 +11,7 @@
 #ifdef MS_RTC_LOGGER_RTP
 #include "RTC/RtcLogger.hpp"
 #endif
-#include <limits> // std::numeric_limits()
+#include <limits> // std::numeric_limits
 
 namespace RTC
 {