+++ worker/test/src/Utils/TestNumber.cpp
@@ -1,6 +1,7 @@
 #include "common.hpp"
 #include "Utils.hpp"
 #include <catch2/catch_test_macros.hpp>
+#include <limits> // std::numeric_limits
 
 using namespace Utils;